fix bug related with path settings

+++ Src/Flask/commands.py
@@ -2,6 +2,8 @@ from models import User
 import click
 from init import app, db
 import shutil
+import os
+
 
 @app.cli.command()  # 注册为命令，可以传入 name 参数来自定义命令
 @click.option('--drop', is_flag=True, help='Create after drop.')  # 设置选项
@@ -11,7 +13,8 @@ def initdb(drop):
         db.drop_all()
     db.create_all()
     click.echo('Initialized database.')  # 输出提示信息
-    shutil.rmtree('./FileStorage/')
+    if os.path.exists('./FileStorage/'):
+        shutil.rmtree('./FileStorage/')
 
 
 @app.cli.command()
